Avance BL

+++ ET/CPU.cs
@@ -5,11 +5,11 @@
         private int id;
         private string descripcion;
         private int idSocket;
-        private float consumoEnergetico;
+        private double consumoEnergetico;
 
         public int Id { get => id; set => id = value; }
         public string Descripcion { get => descripcion; set => descripcion = value; }
         public int IdSocket { get => idSocket; set => idSocket = value; }
-        public float ConsumoEnergetico { get => consumoEnergetico; set => consumoEnergetico = value; }
+        public double ConsumoEnergetico { get => consumoEnergetico; set => consumoEnergetico = value; }
     }
 }
